recouvrement/requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==5.2.10
django-phonenumber-field==8.4.0
mysqlclient==2.2.7
phonenumbers==9.0.22
pillow==12.1.0
python-decouple==3.8
reportlab==4.4.9
sqlparse==0.5.5
tzdata==2025.3
